bot/requirements.txt
APScheduler==3.6.3
bcrypt==4.1.3
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.7
paramiko==3.4.0
psycopg2==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.31.0
six==1.16.0
tornado==6.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
